Refactor data upload
Refactor data upload
